chore: rename the file and folder for about us
