+++ src/app/page.tsx
@@ -2,26 +2,19 @@ import tripData from "./../data/tripData.json";
 import ImageCardGrid from "../components/ImageCardGrid";
 import HeroContent from "@/components/HeroContent";
 import heroData from "./../data/heroData.json";
-import Header from "@/components/Header";
+import AboutContent from "@/components/About";
 
 
 export default function Home() {
   return (
     <>
       <div className="flex items-center justify-center">
-        {/* <h1 className="text-center font-bold text-4xl">Experience Rwanda</h1> */}
       </div>
-      <section id="home">{/* Home content */}</section>
-      <section id="bookings">{/* Bookings content */}</section>
-      <section id="about">{/* About content */}</section>
-      <section id="home">{/* Home content */}</section>
-      <section id="bookings">{/* Bookings content */}</section>
-      <section id="about">{/* About content */}</section>
       <HeroContent
         imageUrl={heroData.imageUrl}
         content={{ title: heroData.title, description: heroData.description }}
       />
-      <Header />
+      <AboutContent />
       <ImageCardGrid
         title={tripData.title}
         description={tripData.description}
